requirements.txt
discord.py
requests
googletrans
aiohttp
beautifulsoup4
imdbpy
wikipedia
flask